+++ R/struct_equiv.R
@@ -5,7 +5,9 @@
 #' @param graph Any class of accepted graph format (see \code{\link{netdiffuseR-graphs}}).
 #' @param v Numeric scalar. Cohesion constant (see details).
 #' @param inf.replace Numeric scalar scalar. Passed to \code{\link[sna:geodist]{sna::geodist}}.
-#' @param ... Further arguments to be passed to \code{\link[sna:geodist]{sna::geodist}}.
+#' @param groupvar Either a character scalar (if \code{graph} is diffnet), or a vector of size \eqn{n}.
+#' @param ... Further arguments to be passed to \code{\link[sna:geodist]{sna::geodist}} (not valid for the print method).
+#' @param x A \code{diffnet_se} class object.
 #' @family statistics
 #' @keywords univar
 #' @details
@@ -35,12 +37,22 @@
 #' the higher will be the influence that the closests alters will have over ego (see
 #' Burt's paper in the reference).
 #'
+#' Structural equivalence can be computed either for the entire graph or by groups
+#' of vertices. When, for example, the user knows before hand that the vertices
+#' are distributed accross separated communities, he can make this explicit to
+#' the function and provide a \code{groupvar} variable that accounts for this.
+#' Hence, when \code{groupvar} is not \code{NULL} the algorithm will compute
+#' structural equivalence within communities as marked by \code{groupvar}.
+#'
+#'
 #' @return If \code{graph} is a static graph, a list with the following elements:
-#' \item{\code{SE}}{Numeric Matrix of size \eqn{n\times n}{n * n} with Structural equivalence}
-#' \item{\code{d}}{Numeric Matrix of size \eqn{n\times n}{n * n} Euclidean distances}
-#' \item{\code{gdis}}{Numeric Matrix of size \eqn{n\times n}{n * n} Normalized geodesic distance}
+#' \item{\code{SE}}{Matrix of size \eqn{n\times n}{n * n} with Structural equivalence}
+#' \item{\code{d}}{Matrix of size \eqn{n\times n}{n * n} Euclidean distances}
+#' \item{\code{gdist}}{Matrix of size \eqn{n\times n}{n * n} Normalized geodesic distance}
 #' In the case of dynamic graph, is a list of size \code{t} in which each element
-#' contains a list as described before.
+#' contains a list as described before. When \code{groupvar} is specified, the
+#' resulting matrices will be of class \code{\link[Matrix:dgCMatrix-class]{dgCMatrix}},
+#' otherwise will be of class \code{\link{matrix}}.
 #'
 #' @references Burt, R. S. (1987). "Social Contagion and Innovation: Cohesion versus
 #' Structural Equivalence". American Journal of Sociology, 92(6), 1287–1335.
@@ -49,58 +61,185 @@
 #' Valente, T. W. (1995). "Network models of the diffusion of innovations" (2nd ed.).
 #' Cresskill N.J.: Hampton Press.
 #' @export
+#'
+#' @examples
+#' # Computing structural equivalence for the fakedata -------------------------
+#' data(fakesurvey)
+#'
+#' # Coercing it into a diffnet object
+#' fakediffnet <- survey_to_diffnet(
+#'    fakesurvey, "id", c("net1", "net2", "net3"), "toa", "group"
+#' )
+#'
+#' # Computing structural equivalence without specifying group
+#' se_all <- struct_equiv(fakediffnet)
+#'
+#' # Notice that pairs of individuals from different communities have
+#' # non-zero values
+#' se_all
+#' se_all[[1]]$SE
+#'
+#' # ... Now specifying a groupvar
+#' se_group <- struct_equiv(fakediffnet, groupvar="group")
+#'
+#' # Notice that pairs of individuals from different communities have
+#' # only zero values.
+#' se_group
+#' se_group[[1]]$SE
+#'
+#'
+#'
 #' @author Vega Yon, Dyal, Hayes & Valente
-struct_equiv <- function(graph, v=1, inf.replace = 0,...) {
-  switch (class(graph),
-    matrix = struct_equiv.matrix(graph, v, inf.replace,...),
-    dgCMatrix = struct_equiv.dgCMatrix(graph, v, inf.replace, ...),
-    array = struct_equiv.array(graph, v, inf.replace, ...),
-    list = struct_equiv.list(graph, v, inf.replace, ...),
-    diffnet = struct_equiv.list(graph$graph, v, inf.replace, ...),
+struct_equiv <- function(graph, v=1, inf.replace = 0, groupvar=NULL,...) {
+
+  # Checking groupvar
+  if ((length(groupvar)==1) && inherits(graph, "diffnet"))
+    groupvar <- graph[[groupvar]]
+
+  output <- switch (class(graph),
+    matrix = struct_equiv.matrix(graph, v, inf.replace, groupvar,...),
+    dgCMatrix = struct_equiv.dgCMatrix(graph, v, inf.replace, groupvar,...),
+    array = struct_equiv.array(graph, v, inf.replace, groupvar,...),
+    list = struct_equiv.list(graph, v, inf.replace, groupvar,...),
+    diffnet = struct_equiv.list(graph$graph, v, inf.replace, groupvar, ...),
     stopifnot_graph(graph)
   )
+
+  structure(output, class="diffnet_se", n = nnodes(graph), nper=nslices(graph),
+            dyn = ifelse(class(graph) %in% c("diffnet", "list", "array"), TRUE, FALSE),
+            inf.replace=inf.replace)
+}
+
+#' @export
+#' @rdname struct_equiv
+print.diffnet_se <- function(x, ...) {
+  dyn <- attr(x, "dyn")
+  cat(paste("Structural equivalence for a", ifelse(dyn, "dynamic", "static"),"graph"),
+      paste(" # nodes :", attr(x, "n")),
+      paste(" # of slices:",attr(x, "nper")),
+      if (dyn) " Access elements via [[nslice]]$ (as a nested list). Available elements are:"
+      else " Access elements via $ (as a list). Available elements are:",
+      "  - SE    : Structural equivalence matrix (n x n)",
+      "  - d     : Euclidean distances matrix (n x n) ",
+      "  - gdist : Structural equivalence matrix (n x n)",
+      sep="\n"
+      )
+
+  return(invisible(x))
 }
 
 
+# Apply per grouping variable
+struct_equiv_by <- function(graph, v, inf.replace, groupvar, ...) {
+  # Checking length of the grouping variable
+  if (length(groupvar) != nvertices(graph))
+    stop("The length of -groupvar-, ",length(groupvar),
+         " elements, must be equal to the number of vertices in the graph, ",
+         nvertices(graph)," vertices.")
+
+  # Checking that is complete
+  test <- which(!complete.cases(groupvar))
+  if (length(test))
+    stop("-groupvar- must not have incomplete cases. Check the following elements:\n\t",
+         paste0(test, collapse=", "), ".")
+
+  # Does the graph has ids?
+  if (!length(rownames(graph))) {
+    rn <- 1:nvertices(graph)
+    dimnames(graph) <- list(rn, rn)
+  }
+
+  # Splitting the data and computing structural equivalence per case
+  G    <- unique(groupvar)
+  n    <- nvertices(graph)
+  SE   <- methods::new("dgCMatrix", Dim=c(n,n), p=rep(0L,n+1L))
+  d    <- SE
+  gdist <- SE
+
+  rn <- NULL
+  N  <- 0L
+  for (g in G) {
+    # Subsetting the graph
+    test <- which(groupvar == g)
+    subg <- graph[test, test, drop=FALSE]
+    rn   <- c(rn, rownames(subg))
+    ng   <- nvertices(subg)
+
+    # Computing SE
+    out <- struct_equiv(subg, v, inf.replace, groupvar=NULL, ...)
+
+    # Appending values
+    index <- (N + 1L):(N + ng)
+    SE[index,index]   <- out$SE
+    d[index,index]    <- out$d
+    gdist[index,index] <- out$SE
+    N <- N + ng
+  }
+
+  # Giving names
+  dimnames(SE) <- list(rn, rn)
+  dimnames(d) <- list(rn, rn)
+  dimnames(gdist) <- list(rn, rn)
+
+  # Ordening output
+  index <- rownames(graph)
+  SE    <- SE[index,index]
+  d     <- d[index, index]
+  gdist  <- gdist[index, index]
+
+  return(list(SE=SE, d=d, gdist=gdist))
+
+}
+
 # @rdname struct_equiv
 # @export
-struct_equiv.matrix <- function(graph, v, inf.replace,...) {
-  geod <- sna::geodist(graph, inf.replace = inf.replace, ...)
-  geod[["gdist"]] <- geod[["gdist"]]/max(geod[["gdist"]], na.rm = TRUE)
-  output <- struct_equiv_cpp(methods::as(geod[["gdist"]], "dgCMatrix"), v)
+struct_equiv.matrix <- function(graph, v, inf.replace, groupvar, ...) {
 
-  # Names
-  rn <- rownames(graph)
-  if (!length(rn)) rn <- 1:nrow(graph)
-  output <- lapply(output, "dimnames<-", value=list(rn, rn))
+  # Running the algorithm
+  if (length(groupvar)) {
+    output <- struct_equiv_by(graph, v, inf.replace, groupvar, ...)
+  } else {
+    geod <- sna::geodist(graph, inf.replace = inf.replace, ...)
+    geod[["gdist"]] <- geod[["gdist"]]/max(geod[["gdist"]], na.rm = TRUE)
+    output <- struct_equiv_cpp(methods::as(geod[["gdist"]], "dgCMatrix"), v)
+
+    # Names
+    rn <- rownames(graph)
+    if (!length(rn)) rn <- 1:nrow(graph)
+    output <- lapply(output, "dimnames<-", value=list(rn, rn))
+  }
 
   return(output)
 }
 
 # @rdname struct_equiv
 # @export
-struct_equiv.dgCMatrix <- function(graph, v, inf.replace,...) {
-  # In order to use the SNA package functions, we need to coerce the graph
-  # Into a -matrix.csc- object,
-  geod <- sna::geodist(methods::as(graph, "matrix.csc"), inf.replace = inf.replace, ...)
-  geod[["gdist"]] <- geod[["gdist"]]/max(geod[["gdist"]], na.rm = TRUE)
-  output <- struct_equiv_cpp(methods::as(geod[["gdist"]], "dgCMatrix"), v)
+struct_equiv.dgCMatrix <- function(graph, v, inf.replace, groupvar, ...) {
 
-  # Names
-  rn <- rownames(graph)
-  if (!length(rn)) rn <- 1:nrow(graph)
-  output <- lapply(output, "dimnames<-", value=list(rn, rn))
+  if (length(groupvar)) {
+    output <- struct_equiv_by(graph, v, inf.replace, groupvar, ...)
+  } else {
+    # In order to use the SNA package functions, we need to coerce the graph
+    # Into a -matrix.csc- object,
+    geod <- sna::geodist(methods::as(graph, "matrix.csc"), inf.replace = inf.replace, ...)
+    geod[["gdist"]] <- geod[["gdist"]]/max(geod[["gdist"]], na.rm = TRUE)
+    output <- struct_equiv_cpp(methods::as(geod[["gdist"]], "dgCMatrix"), v)
 
+    # Names
+    rn <- rownames(graph)
+    if (!length(rn)) rn <- 1:nrow(graph)
+    output <- lapply(output, "dimnames<-", value=list(rn, rn))
+  }
   return(output)
 }
 
 # @rdname struct_equiv
 # @export
-struct_equiv.array <- function(graph, v, inf.replace,...) {
+struct_equiv.array <- function(graph, v, inf.replace, groupvar, ...) {
   t <- dim(graph)[3]
   output <- vector("list", t)
   for(i in 1:t) {
-    output[[i]] <- struct_equiv.matrix(graph[,,i], v, inf.replace, ...)
+    output[[i]] <- struct_equiv.matrix(graph[,,i], v, inf.replace, groupvar, ...)
   }
 
   # Naming
@@ -115,12 +254,13 @@ struct_equiv.array <- function(graph, v, inf.replace,...) {
 
 # @rdname struct_equiv
 # @export
-struct_equiv.list <- function(graph, v, inf.replace, ...) {
+struct_equiv.list <- function(graph, v, inf.replace, groupvar, ...) {
   t <- length(graph)
   n <- nrow(graph[[1]])
   output <- vector("list", t)
   for(i in 1:t)
-    output[[i]] <- struct_equiv.dgCMatrix(methods::as(graph[[i]], "dgCMatrix"), v, inf.replace, ...)
+    output[[i]] <- struct_equiv.dgCMatrix(methods::as(graph[[i]], "dgCMatrix"),
+                                          v, inf.replace, groupvar, ...)
 
   # Naming
   tn <- dimnames(graph)[[3]]